refactor: rename and simplify configuration and initializing process
