+++ npm/react/apps/dev-app/app/root.tsx
@@ -12,20 +12,28 @@ import {
   ClientLoaderFunctionArgs,
   isRouteErrorResponse,
 } from 'react-router';
-
-import { bootstrapApplication } from '@axiomframework/react-core';
-
 import type { Route } from '+/types/routes';
-import { provideEnvironment } from './providers/environment.provider';
 
-provideEnvironment();
+import {
+  initializeApplication,
+  configureLocalization,
+} from '@axiomframework/react-core';
+
+import { configureEnvironment } from './config/index';
+
+function configureApplication() {
+  configureEnvironment();
+  configureLocalization();
+}
+
+configureApplication();
 
 export async function loader({}: LoaderFunctionArgs) {
-  await bootstrapApplication();
+  await initializeApplication();
 }
 
 export async function clientLoader({}: ClientLoaderFunctionArgs) {
-  await bootstrapApplication();
+  await initializeApplication();
 }
 
 export function HydrateFallback() {
@@ -62,7 +70,7 @@ export function Layout({ children }: { children: React.ReactNode }) {
         <Meta />
         <Links />
       </head>
-      <body>
+      <body className="mx-2 my-2">
         {children}
         <ScrollRestoration />
         <Scripts />
